apiquickserve-extdb-plugin/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyodbc==5.1.0
mysql-connector-python==9.3.0
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
